Fix #73 Implement abstract entity class
