+++ src/main/java/com/faforever/api/data/domain/AbstractEntity.java
@@ -1,0 +1,38 @@
+package com.faforever.api.data.domain;
+
+import lombok.Setter;
+
+import javax.persistence.Column;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Inheritance;
+import javax.persistence.InheritanceType;
+import javax.persistence.MappedSuperclass;
+import java.time.OffsetDateTime;
+
+@MappedSuperclass
+@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
+@Setter
+public abstract class AbstractEntity {
+  private int id;
+  private OffsetDateTime createTime;
+  private OffsetDateTime updateTime;
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "id")
+  public int getId() {
+    return id;
+  }
+
+  @Column(name = "create_time")
+  public OffsetDateTime getCreateTime() {
+    return createTime;
+  }
+
+  @Column(name = "update_time")
+  public OffsetDateTime getUpdateTime() {
+    return updateTime;
+  }
+}
