+++ src/test/java/com/faforever/api/clan/ClanServiceTest.java
@@ -176,11 +176,11 @@ public class ClanServiceTest {
     Player leader = new Player();
     leader.setId(3);
 
-    Clan clan = new Clan()
-        .setId(1)
+    Clan clan = (Clan) new Clan()
         .setTag("123")
         .setName("abc")
-        .setLeader(leader);
+        .setLeader(leader)
+        .setId(1);
 
     when(clanRepository.findOne(clan.getId())).thenReturn(clan);
 
@@ -198,11 +198,11 @@ public class ClanServiceTest {
     Player requester = new Player();
     requester.setId(1);
 
-    Clan clan = new Clan()
-        .setId(1)
+    Clan clan = (Clan) new Clan()
         .setTag("123")
         .setName("abc")
-        .setLeader(requester);
+        .setLeader(requester)
+        .setId(1);
 
     when(clanRepository.findOne(clan.getId())).thenReturn(clan);
 
@@ -223,10 +223,10 @@ public class ClanServiceTest {
     Player newMember = new Player();
     newMember.setId(2);
 
-    Clan clan = new Clan()
-        .setId(1)
+    Clan clan = (Clan) new Clan()
         .setTag("123")
-        .setName("abc");
+        .setName("abc")
+        .setId(1);
     clan.setLeader(requester);
 
     FafApiProperties props = new FafApiProperties();
@@ -290,7 +290,7 @@ public class ClanServiceTest {
   @Test
   public void acceptPlayerInvitationTokenInvalidPlayer() throws IOException {
     String stringToken = "1234";
-    Clan clan = new Clan().setId(1);
+    Clan clan = (Clan) new Clan().setId(1);
 
     long expire = System.currentTimeMillis() + 1000 * 3;
     Jwt jwtToken = Mockito.mock(Jwt.class);
@@ -317,7 +317,7 @@ public class ClanServiceTest {
     Player newMember = new Player();
     newMember.setId(2);
 
-    Clan clan = new Clan().setId(1);
+    Clan clan = (Clan) new Clan().setId(1);
 
     Player otherPlayer = new Player();
     otherPlayer.setId(3);
@@ -346,7 +346,7 @@ public class ClanServiceTest {
   public void acceptPlayerInvitationTokenPlayerIAlreadyInAClan() throws IOException {
     String stringToken = "1234";
 
-    Clan clan = new Clan().setId(1);
+    Clan clan = (Clan) new Clan().setId(1);
 
     Player newMember = new Player();
     newMember.setId(2);
@@ -376,7 +376,7 @@ public class ClanServiceTest {
   @Test
   public void acceptPlayerInvitationToken() throws IOException {
     String stringToken = "1234";
-    Clan clan = new Clan().setId(1);
+    Clan clan = (Clan) new Clan().setId(1);
     Player newMember = new Player();
     newMember.setId(2);
     long expire = System.currentTimeMillis() + 1000 * 3;
